0-basic_async_syntax.py

+++ 0x01-python_async_function/0-basic_async_syntax.py
@@ -5,7 +5,7 @@ import asyncio
 import random
 
 
-async def wait_random(max_delay: int) -> int:
+async def wait_random(max_delay: int) -> float:
     """delays for a max delay_max seconds and then returns the delay"""
     delay = random.randint(0, max_delay)
     await asyncio.sleep(delay)
